reconstruct compaction task structure
